# Improve frame grabber performance (and some VS settings)