--- a/src/VHImage.cpp
+++ b/src/VHImage.cpp
@@ -231,7 +231,7 @@ namespace vh {
         VmaAllocationInfo allocInfo;
 
         BufCreateBuffer(allocator, (VkDeviceSize)imageSize, VK_BUFFER_USAGE_TRANSFER_DST_BIT, 
-            VMA_MEMORY_USAGE_CPU_ONLY, VMA_ALLOCATION_CREATE_HOST_ACCESS_SEQUENTIAL_WRITE_BIT | VMA_ALLOCATION_CREATE_MAPPED_BIT,
+            VMA_MEMORY_USAGE_CPU_ONLY, VMA_ALLOCATION_CREATE_HOST_ACCESS_RANDOM_BIT | VMA_ALLOCATION_CREATE_MAPPED_BIT,
             stagingBuffer, stagingBufferAllocation, &allocInfo);
 
 		ImgTransitionImageLayout(device, graphicsQueue, commandPool, image, format,