Improves performance of frame grabber (use HOST_CACHED memory):
* with JPG write: 23 -> 28 FPS
* without JPG write: 90 -> 370 FPS

Sets some convenience settings for VS:
* parallel build
* default startup project to "game"
* working directory for debugger